+++ lib/chewy/railtie.rb
@@ -1,30 +1,23 @@
 module Chewy
   class Railtie < Rails::Railtie
-    module ActionControllerPatch
-      extend ActiveSupport::Concern
-      included do
-        if respond_to?(:prepend_around_action)
-          prepend_around_action :setup_chewy_strategy
-        else
-          prepend_around_filter :setup_chewy_strategy
-        end
+    class RequestStrategy
+      def initialize(app)
+        @app = app
       end
 
-    private
-
-      def setup_chewy_strategy
-        Chewy.strategy(:atomic) { yield }
+      def call(env)
+        Chewy.strategy(:atomic) { @app.call(env) }
       end
     end
 
-    module ActiveRecordMigrationPatch
+    module MigrationStrategy
       extend ActiveSupport::Concern
       included do
         alias_method_chain :migrate, :chewy
       end
 
-      def migrate_with_chewy(*_)
-        Chewy.strategy(:bypass) { migrate_without_chewy(*_) }
+      def migrate_with_chewy(*args)
+        Chewy.strategy(:bypass) { migrate_without_chewy(*args) }
       end
     end
 
@@ -41,18 +34,18 @@ module Chewy
       end
     end
 
-    initializer 'chewy.migrations_patch' do
+    initializer 'chewy.logger', after: 'active_record.logger' do
+      ActiveSupport.on_load(:active_record)  { Chewy.logger ||= ActiveRecord::Base.logger }
+    end
+
+    initializer 'chewy.migration_strategy' do
       ActiveSupport.on_load(:active_record) do
-        ActiveRecord::Migration.send(:include, ActiveRecordMigrationPatch)
+        ActiveRecord::Migration.send(:include, MigrationStrategy)
       end
     end
 
-    initializer 'chewy.action_wrapper' do
-      ActiveSupport.on_load(:action_controller) { include ActionControllerPatch }
-    end
-
-    initializer 'chewy.logger', after: 'active_record.logger' do
-      ActiveSupport.on_load(:active_record)  { Chewy.logger ||= ActiveRecord::Base.logger }
+    initializer 'chewy.request_strategy' do |app|
+      app.config.middleware.insert_after(Rack::Runtime, RequestStrategy)
     end
 
     initializer 'chewy.add_app_chewy_path' do |app|